Fix volume store name, refactor test support.
